+++ src/app/shared/components/add-to-cart-btn/add-to-cart-btn.component.spec.ts
@@ -1,0 +1,25 @@
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { AddToCartBtnComponent } from './add-to-cart-btn.component';
+
+describe('AddToCartBtnComponent', () => {
+  let component: AddToCartBtnComponent;
+  let fixture: ComponentFixture<AddToCartBtnComponent>;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [ AddToCartBtnComponent ]
+    })
+    .compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(AddToCartBtnComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});