+++ changelog.md
@@ -2,3 +2,9 @@
 "Task 2. Components" was completed. All subtasks were done. The additional functionality is changing the background of the "Add to cart" buttons on the main product list onmouseover.
 "Task 3. Services and DI" was completed. All subtasks were done.
 "Task 4. Pipes" was completed. The first pipe is in the "About" block next to the "Task 4.1." pointer. The second task is in the cart. Please add products to the cart, then open the cart and apply sorting. 
+
+Questions
+1. why this does not work? (ProductComponent - Constructor)
+   this.productService.getProduct(productId).toPromise().then(val => console.log(val));
+
+2. How to pass additional CSS class to a child element? The class should be applied not to the element itself, but to the specific input inside the child html.